* commit 'b14c09a0a577bd70621e4ad482d5357a503fe8a1':
  bug
  bug

+++ fbi_conf/src/com/fr/bi/conf/data/source/ExcelTableSource.java
@@ -98,7 +98,7 @@ public class ExcelTableSource extends AbstractTableSource implements JSONTransfo
             @Override
             public void actionPerformed(BIDataValue v) {
                 Object resValue = null;
-                long i = v.getRow();
+                int i = v.getRow();
                 int j = v.getCol();
                 Object value = v.getValue();
                 switch (columns[j].getFieldType()) {
